+++ cli/src/cmd/project/init_test.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package apiproject
+package project
 
 import (
 	"os"
