+++ cli/src/cmd/root.go
@@ -26,7 +26,7 @@ import (
 	"github.com/wso2/api-platform/cli/cmd/devportal"
 	"github.com/wso2/api-platform/cli/cmd/gateway"
 	"github.com/wso2/api-platform/cli/cmd/platform"
-	"github.com/wso2/api-platform/cli/cmd/apiproject"
+	"github.com/wso2/api-platform/cli/cmd/project"
 	"github.com/wso2/api-platform/cli/utils"
 )
 
@@ -70,7 +70,7 @@ func init() {
 	rootCmd.AddCommand(devportal.DevPortalCmd)
 	rootCmd.AddCommand(gateway.GatewayCmd)
 	rootCmd.AddCommand(platform.PlatformCmd)
-	rootCmd.AddCommand(apiproject.ApiProjectCmd)
+	rootCmd.AddCommand(project.ProjectCmd)
 	rootCmd.AddCommand(versionCmd)
 }
